refactor: getBranchCommits
