+++ electron/app/services/git/handlers/getBranchCommits.ts
@@ -1,0 +1,19 @@
+import {orderBy} from 'lodash';
+import simpleGit from 'simple-git';
+
+import {GitBranchCommitsParams, GitBranchCommitsResult} from '@shared/ipc/git';
+
+export async function getBranchCommits({
+  branchName,
+  localPath,
+}: GitBranchCommitsParams): Promise<GitBranchCommitsResult> {
+  const git = simpleGit({baseDir: localPath});
+
+  try {
+    const commits = [...(await git.log({[branchName]: null})).all];
+
+    return orderBy(commits, ['date'], ['desc']);
+  } catch (e: any) {
+    throw new Error(e.message);
+  }
+}
