+++ electron/app/services/git/handlers/index.ts
@@ -1,6 +1,7 @@
 export * from './checkoutGitBranch';
 export * from './cloneGitRepo';
 export * from './getAheadBehindCommitsCount';
+export * from './getBranchCommits';
 export * from './getGitRemotePath';
 export * from './isFolderGitRepo';
 export * from './isGitInstalled';
