+++ src/redux/git/service/checkoutGitBranch.ts
@@ -1,6 +1,6 @@
 import {invokeIpc} from '@utils/ipc';
 
-import {GitCheckoutBranchParams} from '@shared/ipc';
+import {GitCheckoutBranchParams} from '@shared/ipc/git';
 
 /**
  * Example usage:
