+++ src/redux/git/service/pullChanges.ts
@@ -1,6 +1,6 @@
 import {invokeIpc} from '@utils/ipc';
 
-import {GitPathParams} from '@shared/ipc';
+import {GitPathParams} from '@shared/ipc/git';
 
 /**
  * Example usage:
